Add MBE280 files.

+++ avr/SothCoreATXMEGA32E5/SothCore/BME280.h
@@ -13,7 +13,7 @@
 #define BME280_TRIM_DATA_LENGTH 32
 
 void initBME280();
-void readTemperaturePressureHumidity(void (*func)(double temperature, double pressure, double humidity));
+void readTemperaturePressureHumidity(void (*func)(float temperature, float pressure, float humidity));
 
 
 #endif /* BME280_H_ */
